+++ com.unity.netcode.gameobjects/CHANGELOG.md
@@ -10,11 +10,18 @@ Additional documentation and release notes are available at [Multiplayer Documen
 
 ### Added
 
+### Removed
+
+- Removed `FixedQueue` (#1398)
+- Removed `StreamExtensions` (#1398)
+- Removed `TypeExtensions` (#1398)
+
 ### Fixed
 
 - Fixed in-scene NetworkObjects that are moved into the DDOL scene not getting restored to their original active state (enabled/disabled) after a full scene transition (#1354)
 - Fixed invalid IL code being generated when using `this` instead of `this ref` for the FastBufferReader/FastBufferWriter parameter of an extension method. (#1393)
 - Fixed an issue where if you are running as a server (not host) the LoadEventCompleted and UnloadEventCompleted events would fire early by the NetworkSceneManager (#1379)
+- Fixed a runtime error when sending an array of an INetworkSerializable type that's implemented as a struct (#1402)
 
 ### Changed